cli: Restrict the privileges of CLI kube credentials
We now only allow it to get logs and exec. Note that it's still possible
to exec into the sandbox controller if you knew it existed, but even
attackers would still need to break out of their namespace. Because we
are going to get rid of connecting directly to Kubernetes from the CLI
anyways, this is good enough for now.