up: Fix bug with status stream cancellation
Before, we stopped watching the status if the grpc stream was cancelled,
even if it wasn't cancelled by the status watcher (e.g. it's possible
for another piece of code to close the parent ClusterManager
connection).

This commit makes it so we always retry getting the status until all
pods are running.