volumes: Make volume identifier unique for relative bind volumes
Before, if the user switched to a Docker Compose file that was in a
different directory, but had a bind volume with the same relative path
(e.g. `.`), Blimp would use the same host directory for the volume. This
would lead to the other volume's files getting synced back to the local
filesystem.