wait: Properly block container boot on initial volume sync
The old code that blocked the container boot didn't work because of a
number of incorrect assumptions about the Syncthing API.

Because Syncthing doesn't have a concept of "an initial sync", this
patch makes it so we determine when the sync is complete by having the
CLI publish a hash representing the expected contents of the volume. We
then delay starting the user's container until the remote hash matches
the expected hash.