network: Don't allow inter-namespace traffic
Before, the networks for each namespace weren't actually namespace, so
if you guessed the IP of another customer's container on the network,
you could connect to it.

This commit adds a NetworkPolicy, to each namespace that disallows
connections from other namespaces. It requires a CNI to enforce the
NetworkPolicy. Because we use GKE, we just enable Calico when creating
the cluster.