up: Don't rebuild images by default
This commit changes the behavior of `up` so that it only rebuilds images
if they `--build` flag is set, or it's the first time running `blimp up`
for the given compose file.

This mimics the behavior of Docker Compose, and greatly speeds up boot
times, since inconsequential changes won't trigger a full image rebuild.