sync: Properly calculate hash for broken symlinks
Before, we would always try to open the symlink's destination, which
would fail if the symlink was broken. We now only account for the
destination of the symlink, but don't actually read its contents. If the
symlink's destination is synced as well, the hash of that file is
sufficient to make sure the sync is consistent.