login: Rewrite flow for getting credentials
Before, the CLI obtained the login token after logging in from the
browser by spinning up a local web server, and setting the OAuth
redirect URL to localhost.

This was brittle, since it depended so heavily on the configuration of
people's local system. For example, we ran into people who had other
processes listening on the same port locally, and people with odd DNS
setups.

This commit makes it so the OAuth token is retrieved server-side, and
pushed back to the CLI.