sync: Follow symlinks for bind volume roots when hashing
Before, we would wouldn't properly hash the contents of bind volumes
when the bind volume was a symlink. For example, the following directory
structure would be stuck in "Syncing" when run from the `symlinked-repo`
directory:

```
my-repo
symlinked-repo -> my-repo
```

```
volumes:
- '.:/vol'
```

This was because Syncthing would (correctly) sync the contents of
`my-repo` into the container, but the local hashing code wouldn't take
the contents of `my-repo` into account when hashing.

Note that we still do NOT follow symlinks for files _within_ a volume.