volumes: Support mounting individual files
This commit rewrites our syncthing approach to allow mounting individual
files. Before, we only supported mounting directories.

The syncthing pod now syncs all of the files into a single Kubernetes
volume. The structure of the volume mirrors the structure of the user's
filesystem, and the cluster controller just projects the appropriate
pieces of the volume into pods via VolumeMounts.

The CLI is responsible for ensuring the files referenced by volumes are
synced. For example, when mounting an individual file, the CLI actually
syncs the parent directory, and uses a stignore file to ensure that only
the file that's being mounted is synced.