status: Handle image pull states better
We now tell the user when a pod is stuck in pending because the image
failed to pull, and when the image is pulling.

To achieve this, the status API now has a Phase field, which has a
limited set of valid options, and a Message field, which can contain
arbitrary additional information.