cluster: Set a maximum number of concurrent sandboxes
If we already have 100 sandboxes and someone new tries to boot, we will politely
reject them and tell them to try again later. To count how many sandboxes are
currently booted, we add a label blimp.sandbox=true to all sandbox namespaces,
and just count the number of namespaces with this label. The maximum number of
sandboxes can be overridden with the MAX_SANDBOXES environment variable.