init: Gracefully handle connection failures to Node controller
Before, if the init container failed to connect to the node controller,
the container could OOM, since it didn't clean up its connection, and
retried in a tight loop.

This commit improves the behavior when we end up in a bad state, but
doesn't fix the underlying reason why the container can fail to connect
to the node controller, which is that if the node gets removed from the
cluster, the IP will no longer exist.