logs: Check that service has started
Before, running `blimp logs` on a service that didn't exist or hadn't booted
would give a nasty error about the pod not being found. Now, we notice this
situation and give a more graceful message.

This requires a broader check than exec/ssh/cp, since we want to be able to
check logs from pods that have exited, or are pending in CrashLoopBackoff. To
accomplish that, we generalize the CheckServiceRunning function to take a
predicate that allows arbitrary checks on the ServiceStatus.

This commit also changes the default check used in exec/ssh/cp to allow service
that are UNHEALTHY, since I realized that should probably be allowed.