syncthing: Avoid generate conflicting ignore rules
When an Stignore is generated for a Mount that has Includes that are
nested relative to each other, we generate an Stignore that ignores the
parent Include. For example, a Mount with the following includes:
- foo/
- foo/bar

Generates the following stignore in order to include `foo/bar`:
```
!foo/bar
foo/bar/
!foo/bar
foo/
!foo
```

However, the 4th line (`foo/`) incorrectly ignores all files within
`foo/`, even though the Mount declares that it should be synced.

This commit avoids this issue by fixing the bug that generated Mounts
with nested Includes. The code now recognizes when an Include is
unnecessary, and avoids adding it.