node: Autoupdate node controller versions
Before, if the cluster manager was deployed with a new version of the
node controller, it wouldn't upgrade any existing node controllers. This
commit makes it so the cluster manager resyncs the node controller's pod
spec, even if the node controller already exists. This way, if the pod
spec changes (e.g. if the image changes), the node controller will be
updated.

As a side benefit of this change, which converts the code to use a
separate workqueue, we now create node controllers in parallel, and
retry deployments after failure.