vcp: Preserve permissions from container when copying
This commit modifies our volume initialization logic to copy the
directory's permissions from the image into the volume. This fixes a bug
where volumes couldn't be used with containers that don't run as root,
such as Redis. Before, we would always create the volume as root, so the
container wouldn't be able to write into the volume, and would crash.