cluster: Allow containers to share built images
If `build` and `image` are both specified by a service in docker-compose.yml,
Docker Compose will build the given context and then tag it with the given image
reference. This allows you to use this image in other services in the same
docker-compose.yml. We emulate this behavior by noticing when a built image is
should be tagged, and telling other services using this tag to use the built
image instead.