deploy: Increase limits to allow elasticsearch to run
Elasticsearch (in certain configurtaions) needs a high memlock limit, and it
needs the sysctl vm.max_map_count to be at least 262144. We set both of these at
the node level. Ideally, we should be able to set the ulimit at the container
level (to match how it is set in Docker Compose), but Kubernetes doesn't
currently support this.