kube: Fix race in deletion observation
`DeletePod` blocks until we've observed that the pod has been deleted.
We implement this by creating a watch for the pod's namespace, which
triggers checks to see if the pod has been deleted. However, Watch
returns all events _after_ the watch was initiated. Therefore, if the
deletion completed before the watch began, and no other changes happened
in the namespace, the deletion code would hang indefinitely.

We now do an initial poll after initiating the watch. In case we miss
events from the watch, we also poll every 30 seconds.