build: Pre-push base images
Before, image pushes could be very slow, especially on bad internet connections.
Now, blimp asks the cluster-controller to pre-push the base image as defined in
the Dockerfile for any custom imges. Then, when the custom image is pushed,
Docker will recognize that the layers from the base image are already present on
the registry, and will not push these layers.

Co-authored-by: Kevin Lin <kevin@kelda.io>