registry: Locally cache JSON Web Key Set to disk
Since the registry auth is invoked as a separate process each time, the JWKS
does not get cached in local memory. This creates an alterntaive implementation
of oidc.KeySet that is similar to oidc.remoteKeySet, but simpler and with a
cache on disk instead of in memory. This allows us to handle more authentication
requests at once without getting ratelimited by auth0, which could happen if a
lot of images were pushed at the same time.