dns: Support internal hostnames with TLDs
Example uses reprise-development.com.

Note that this introduces a possible race where if a hostname exists
both internally and externally, Blimp can respond with the external IP
if it hasn't learned about the pod with the internal IP at the time of
the request. However, this is still an improvement from the previous
code, where we didn't attempt to check the hostname internally at all.