wait: Ensure container stays up for at least 5 seconds
A bug in Docker causes short-lived containers to appear like they
failed, even if they exited cleanly:
https://github.com/opencontainers/runc/issues/2183

This can happen if the wait spec is already satisfied when this
container starts.

To work around this, we add an artificial sleep to make sure the
container stays up for at least 5 seconds.

This fixes the "OCI runtime start failed: container process is already
dead: unknown" error.