syncthing: Avoid syncing directories masked by native volumes
This allows you to specify your working directory as a bind volume, but specify
./node_modules as a native volume, to avoid having to sync all the changes to
that directory back and forth from your laptop to your sandbox.

The logic is currently best-effort. It notices common cases that allow
subdirectories to be ignored during syncing, but it does not guarantee that it
will notice if a directory can be ignored. This is to avoid ridiculous code
complexity and/or a total rewrite of how we collapse syncthing mounts.