cluster: Deploy NetworkPolicy even if namespace already exists
Before, we didn't bother trying to deploy some of the objects within a
namespace if it already existed, since we assumed that the objects would
already exist.

However, an upcoming commit will add a PVC that needs to be in each
namespace. This commit makes it easier to make it so that the PVC will
be deployed into existing namespaces.