volume: Make volumes persist across `blimp down`s
We now back volumes with PersistentVolumes rather than HostPath volumes,
and don't clear the volume until `blimp down --volumes` is run. Moving
to PersistentVolumes is required to achieve this so that namespaces can
be scheduled to a different node after `blimp down`.

Note that this patch removes the finalizer handler for the old host path
volumes. To make the upgrade simple, we should terminate all existing
namespaces prior to upgrading.