status: Provide tip when container is in CreateContainerError
It seems like the Kubelet can sometimes get into a bad state and fail to
create the container, and never retry. When this happens, we now give a
more informative message to the user saying that blimp probably won't
recover.

Because this error is difficult to reproduce, I didn't confirm that the
message is shown in live testing. I did confirm that the normal statuses
still show up.