build: Add command for building individual services
Before, the only way of triggering a rebuild was the `--build` flag for
`blimp up`, which would rebuild _all_ services. This could be very time
consuming if there were lots of services.

This commit adds `blimp build`, which takes a list of services to build
as arguments, and only builds those services.