up: Protect against unresponsive Docker daemon
This should help make it clear when Docker is causing issues instead of us. We
set short (30s) timeouts on queries that should be quick and do not require a
network connection. We also print messages indicating what we're doing before
launching into any potentially long-running Docker operations such as building
or pushing.