dns: Only update table when IP changes
Before, we would see high CPU usage (20%) on the DNS server, even if no
requests were being made. This is because the DNS server would hit the
Kubernetes API every time a pod in the namespace changed. This would
happen regularly when a pod used healthchecks because Kubernetes updates
the pod status after running each probe.

This commit fixes this by only updating the DNS table if the pod's IP
changes. It uses a Kubernetes informer to watch pods, which has the
additional benefit that we don't hit the Kubernetes API when syncing
the table.