logs: When following, only exit if all containers exit
Before, any time the logs stream ended, we would treat it as if the underlying
container had died. However, network transport issues can cause EOF or other
errors that are unrelated to the status of the pod. In these cases, we shouldn't
exit, we should try to reconnect to the logs stream.

To accomplish this, we watch the status of the sandbox to notice when pods die
and restart. If we are following logs, we ONLY exit if we have gotten status
updates that show that all the services we are logging have exited, or the
sandbox has terminated. Otherwise, in the case of container restarts, service
restarts, network issues, or other potential failure cases, we continually try
to reconnect.