up: Don't push base image tags
It turns out that pushing the exact tag for the base image is only necessary for
cross-repository push mounts, but since we are pushing to the `:base` tag in the
same repository, we don't actually need to do this. This commit removes the code
to do this and cleans up some code that was only necessary so that we could do
this.