cluster: Abort CreateSandbox if we find errors in the compose file
Some users had DependsOn that referred to services that were not defined in the
Docker Compose file, but we did not notice and would gladly create spin up the
sandbox. Then these containers would be stuck waiting on their nonexistent
dependencies forever. This adds a code path to notice this issue and potentially
future similar issues.