cluster: Taint nodes until their corresponding LB has DNS
On EKS, load balancers are hostname-based, but the DNS for the hostname doesn't
become available for 2-3 minutes after the hostname is added to the Service.
This commit taints the node while waiting for the DNS records to become
available, to make sure that sandboxes aren't scheduled onto a node that the CLI
won't be able to connect to.