up: Cancel all GUI output when exiting
Before, if you hit ctrl-c while the sandbox was still starting, you would
continue to get status updates while the sandbox was being destroyed. Now, we
cancel logs and status updates as when you hit ctrl-c.

Logs already had its own ctrl-c handler, so there is no meaningful change there,
but we still pass in a context for consistency, in case we introduce another way
to cancel `up` in the future.