link-proxy: Add HTTP server to handle requests for exposed services
The link-proxy is a new component that will handle incoming requests for exposed
services (all requests to *.blimp.dev). TLS should be handled in front of this
server and traffic should be reverse proxied.

This implementation is a prototype that works but is currently unoptimized and
may handle errors and edge cases poorly.