syncthing: Don't write in-use .stignore
Writing (or maybe replacing) a file that is in-use on Windows will
give a permission error. We should only write to the file if it's
missing for some reason. In practice, on Windows this shouldn't be
possible since you shouldn't be able to delete the file as long as
syncthing has it open.