cluster: Don't allow pre-pushes to other repositories
This mitigates a vulnerability that would allow users to make pre-push requests
to arbitrary docker repositories. This is unlikely to be very easily exploitable
in the real world, since the registry auth is hardcoded to use your blimp auth
token, but still could be exploited in some obscure edge cases.