dockercompose: Implement "extends" by upgrading compose-go version
Compose-go now implements "extends", so this commit upgrades the CLI to
use the newest version.

Note that compose-spec/compose-go also recently changed their API to
return a `types.Project` rather than `types.Config`, so this commit also
transitions to using the new type. Fortunately, the type is fully backwards
compatible, so the Compose serialization to and from the manager cluster
doesn't have to change.