extends: Normalize paths when using "extends"
Before, if a compose file extended a compose file in another directory,
the parsed result would resolve file references in the base file as
relative to the importing file. For example, context for the following
compose files would resolve to "./importer" rather than "./base".

./importer/docker-compose.yml:
```
services:
  web:
    extends:
      file: ../base/docker-compose.yml
      service: web
```

./base/docker-compose.yml:
```
services:
  web:
    build: .
```

This issue was fixed in the `github.com/kelda/compose-go` package. This
commit makes it so we use the version with the fix.