vcp: Resolve source symlinks
Before, if the path within the image was a symlink, we would just copy
the symlink, rather than the contents within the directory it points to.
This would cause Kubernetes to fail to start the runtime container
because it'd see that the volume mount already exists, but is not a
directory (it'd be a symlink).