analytics: Remove version check
We set the version to `latest` for the CLI pods running in the cluster,
so their analytics weren't being published.

Note that we still discard analytics in development because the cluster
manager still has the version check. So the analytics will just get
discarded once they've been received by the manager.