expose: Allow multiple service/port combinations to be exposed
This changes the expose subdomain format to allow for multiple services or
multiple ports on the same service to be exposed. It also changes the subdomains
to include a portion that is randomly generated, so that the subdomain cannot be
guessed or enumerated even if you know the namespace, service, and port you want
to connect to.