auth: Add support for cluster authorization
This updates to use a new version of the protobuf that uses generic auth message
which includes an optional cluster_auth field. This is a symmetric token that
can be specified via an environment variable to bar access to users without the
token.

The approach taken in this PR generalizes much of the authentication logic so
that future updates to authentication/authorization ideally will not have to
touch so many places.