auth: Add support for cluster authorization
This updates the protobuf to use a generic auth message which includes an
optional cluster_auth field. The "old_token" field is preserved in the protobuf,
so that the backend can still read the token from old clients that send that
field instead of the new auth field. To use the new cluster_auth token, simply
set the "cluster_token" field in ~/.blimp.yaml.

This also generalizes some areas of the code that rely on auth being a single
string, such as the registry credential for the docker registry.